Wow so i found some bggy stuff in my old code -_- thought id be done 4 hrs ago, but im basically done now. No need for menu bar anymore so... Close #7. I created a crappy, but functional how to play section. Close #8. I changed my bad boss to my friends face being the last thing. will update readme and other file later. Close #11

+++ background.cpp
@@ -43,12 +43,12 @@ void Background::move()
 	}
 }
 
-int Background::getX()
+float Background::getX()
 {
 	return x;
 }
 
-int Background::getY()
+float Background::getY()
 {
 	return y;
 }
